pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = ""
authors = [
    {name = "zoeyy89",email = "zoeychien0910@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)"
]

[tool.poetry]
packages = [{include = "myproject", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
